refactor: split and move main file to cmd and move logic to internal
Split HTTP and gRPC servers' bootstrapping and running into separate commands, placed under the cmd folder.
Move logic to the internal folder.
Update imports, Makefile commands and gitignore.
