+++ Makefile
@@ -13,15 +13,18 @@ init:
 # generate proto files
 proto:
 	rm -rf pb doc/swagger && \
-	mkdir pb && mkdir -p doc/swagger && \
-	protoc --proto_path=proto --go_out=pb --go_opt=paths=source_relative \
-    --go-grpc_out=pb --go-grpc_opt=paths=source_relative \
-	--grpc-gateway_out=pb --grpc-gateway_opt=paths=source_relative \
+	mkdir -p internal/pb && mkdir -p doc/swagger && \
+	protoc --proto_path=proto --go_out=internal/pb --go_opt=paths=source_relative \
+	--go-grpc_out=internal/pb --go-grpc_opt=paths=source_relative \
+	--grpc-gateway_out=internal/pb --grpc-gateway_opt=paths=source_relative \
 	--openapiv2_out=doc/swagger --openapiv2_opt=allow_merge=true,merge_file_name=gatehub-data-api \
+	--experimental_allow_proto3_optional \
     proto/*.proto \
 	&& \
-	protoc-go-inject-tag -input="./pb/*.pb.go"
+	protoc-go-inject-tag -input="./internal/pb/*.pb.go"
 
 .PHONY: start
-start: 
-	go run main.go
+start:
+	go run cmd/grpc/main.go & \
+	go run cmd/http/main.go & \
+	wait
