+++ internal/gapi/rpc_get_account_balance_changes.go
@@ -2,13 +2,15 @@ package gapi
 
 import (
 	"context"
-	"github.com/GateHubNet/data-api-go/models"
-	"github.com/GateHubNet/data-api-go/pb"
-	"github.com/GateHubNet/data-api-go/util"
+
 	"github.com/rs/zerolog/log"
 	"github.com/scylladb/gocqlx/v2/qb"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
+
+	"github.com/GateHubNet/data-api-go/internal/models"
+	"github.com/GateHubNet/data-api-go/internal/pb"
+	"github.com/GateHubNet/data-api-go/internal/util"
 )
 
 func (server *Server) accountBalanceChangesResponseHelper(ctx context.Context, balanceChanges []*pb.BalanceChanges) *pb.GetAccountBalanceChangesResponse {
