+++ internal/gapi/rpc_get_ledger.go
@@ -7,14 +7,15 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/GateHubNet/data-api-go/models"
-	"github.com/GateHubNet/data-api-go/pb"
-	"github.com/GateHubNet/data-api-go/util"
 	"github.com/rs/zerolog/log"
 	"github.com/scylladb/gocqlx/v2/qb"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 	"google.golang.org/protobuf/proto"
+
+	"github.com/GateHubNet/data-api-go/internal/models"
+	"github.com/GateHubNet/data-api-go/internal/pb"
+	"github.com/GateHubNet/data-api-go/internal/util"
 )
 
 func (server *Server) getLedgerByIndex(ledgerIndex int64) (*pb.Ledger, error) {
