--- main.go
@@ -1,117 +1,0 @@
-package main
-
-import (
-	"context"
-	"net"
-	"net/http"
-	"os"
-	"strconv"
-
-	"google.golang.org/protobuf/proto"
-
-	"github.com/GateHubNet/data-api-go/gapi"
-	"github.com/GateHubNet/data-api-go/pb"
-	"github.com/GateHubNet/data-api-go/util"
-	"github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
-	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/reflection"
-)
-
-func httpResponseModifier(ctx context.Context, w http.ResponseWriter, p proto.Message) error {
-	md, ok := runtime.ServerMetadataFromContext(ctx)
-	if !ok {
-		return nil
-	}
-
-	// set http status code
-	if vals := md.HeaderMD.Get("x-http-code"); len(vals) > 0 {
-		code, err := strconv.Atoi(vals[0])
-		if err != nil {
-			return err
-		}
-		// delete the headers to not expose any grpc-metadata in http response
-		delete(md.HeaderMD, "x-http-code")
-		delete(w.Header(), "Grpc-Metadata-X-Http-Code")
-		w.WriteHeader(code)
-	}
-
-	return nil
-}
-
-func main() {
-	config, err := util.LoadConfig(".")
-	if err != nil {
-		log.Fatal().Err(err).Msg("cannot load config")
-	}
-
-	if config.Environment == "development" {
-		log.Logger = log.Output(zerolog.ConsoleWriter{Out: os.Stderr})
-	}
-
-	go runGatewayServer(config)
-	runGrpcServer(config)
-}
-
-func runGrpcServer(config util.Config) {
-	server, err := gapi.NewServer(config)
-	if err != nil {
-		log.Fatal().Err(err).Msg("cannot create server")
-	}
-
-	defer server.Store.Close()
-	//TODO: handle error ... this actually doesnt matter because if this functions ends, server should also end
-	defer server.Cache.Close()
-
-	grpcLogger := grpc.UnaryInterceptor(gapi.GrpcLogger)
-	grpcServer := grpc.NewServer(grpcLogger)
-	pb.RegisterGateHubDataAPIServer(grpcServer, server)
-	reflection.Register(grpcServer)
-
-	listener, err := net.Listen("tcp", config.GRPCServerAddress)
-	if err != nil {
-		log.Fatal().Err(err).Msg("cannot create grpc listener")
-	}
-
-	log.Info().Msgf("start gRPC server at %s", listener.Addr().String())
-	err = grpcServer.Serve(listener)
-	if err != nil {
-		log.Fatal().Err(err).Msg("cannot start gRPC server")
-	}
-}
-
-func runGatewayServer(config util.Config) {
-	server, err := gapi.NewServer(config)
-	if err != nil {
-		log.Fatal().Err(err).Msg("cannot create server")
-	}
-
-	defer server.Store.Close()
-	//TODO: handle error ... this actually doesnt matter because if this functions ends, server should also end
-	defer server.Cache.Close()
-
-	grpcMux := runtime.NewServeMux(runtime.WithForwardResponseOption(httpResponseModifier))
-
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-
-	err = pb.RegisterGateHubDataAPIHandlerServer(ctx, grpcMux, server)
-	if err != nil {
-		log.Fatal().Err(err).Msg("cannot register handler server")
-	}
-
-	mux := http.NewServeMux()
-	mux.Handle("/", grpcMux)
-
-	listener, err := net.Listen("tcp", config.HTTPServerAddress)
-	if err != nil {
-		log.Fatal().Err(err).Msg("cannot create HTTP listener")
-	}
-
-	log.Info().Msgf("start HTTP server at %s", listener.Addr().String())
-	err = http.Serve(listener, mux)
-	if err != nil {
-		log.Fatal().Err(err).Msg("cannot start HTTP server")
-	}
-}
